dashboard/pyproject.toml
[tool.poetry]
name = "dashboard"
version = "0.1.0"
description = ""
authors = ["James Wood <s4638772@student.uq.edu.au>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
libsass = "^0.23.0"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"
dlib = "^19.24.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
